LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/infrastructure/urls/resolving-urls/query-encoding/utf-8_include=websocket-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo-query?%C3%A5' failed: WebSocket is closed before the connection is established.

PASS WebSocket constructor
PASS WebSocket#url
